moved schemas to adapters and services to handlers
